feat(UI-1019): events drawer - refactor and optimize
