+++ src/contexts/index.ts
@@ -1,1 +1,3 @@
+export { EventsDrawerProvider } from "@src/contexts/eventsDrawer/eventsDrawerProvider";
+export { useEventsDrawer } from "@src/contexts/eventsDrawer/useEventsDrawer";
 export { PopoverContext, usePopoverContext, PopoverListContext, usePopoverListContext } from "@src/contexts/usePopover";
